+++ srcs/ft_printhex.c
@@ -6,13 +6,13 @@
 /*   By: btan <btan@student.42.singapore.sg>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/24 11:26:37 by btan              #+#    #+#             */
-/*   Updated: 2023/09/26 08:28:27 by btan             ###   ########.fr       */
+/*   Updated: 2023/09/26 09:25:42 by btan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/ft_printf.h"
 
-int	ft_printhex(unsigned int n, char *base)
+int	ft_printhex(unsigned long n, char *base)
 {
 	int	count;
